# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dwizzyBRAIN API",
    "version": "1.8.0",
    "description": "Standardized read API for market snapshots, OHLCV candles, aggregated tickers, arbitrage signals, quant pattern search, DeFi registry reads, news feeds, and auth/session endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/v1/auth/discord/start": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Redirect to Discord OAuth",
        "operationId": "redirectDiscordOAuth",
        "responses": {
          "302": {
            "description": "Redirect to Discord authorization URL"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create Discord OAuth start payload",
        "operationId": "createDiscordOAuthStart",
        "responses": {
          "200": {
            "description": "Discord authorization URL and state payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStartResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
    "/v1/auth/discord/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Complete Discord OAuth callback",
        "operationId": "completeDiscordOAuth",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated session response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
    "/v1/auth/web3/nonce": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a Web3 login nonce",
        "operationId": "createWeb3LoginNonce",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Web3NonceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nonce challenge payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Web3NonceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
    "/v1/auth/web3/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify a Web3 signature",
        "operationId": "verifyWeb3LoginSignature",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Web3VerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Rotate refresh token",
        "operationId": "refreshAuthSession",
        "responses": {
          "200": {
            "description": "Rotated access and refresh token response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke the current auth session",
        "operationId": "logoutAuthSession",
        "responses": {
          "200": {
            "description": "Logout confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLogoutResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get the current authenticated session",
        "operationId": "getCurrentAuthSession",
        "responses": {
          "200": {
            "description": "Current authenticated user and session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequestError": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "UnauthorizedError": {
        "description": "Authentication required or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServiceUnavailableError": {
        "description": "Authentication service is unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "AuthStartData": {
        "type": "object",
        "required": [
          "authorization_url",
          "state",
          "state_expires_at"
        ],
        "properties": {
          "authorization_url": {
            "type": "string",
            "format": "uri"
          },
          "state": {
            "type": "string"
          },
          "state_expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthStartResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthStartData"
          }
        }
      },
      "Web3NonceRequest": {
        "type": "object",
        "required": [
          "wallet_address"
        ],
        "properties": {
          "wallet_address": {
            "type": "string",
            "example": "0x1234567890abcdef1234567890abcdef12345678"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "login"
            ],
            "default": "login"
          }
        }
      },
      "Web3NonceData": {
        "type": "object",
        "required": [
          "wallet_address",
          "purpose",
          "nonce",
          "challenge",
          "expires_at"
        ],
        "properties": {
          "wallet_address": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "nonce": {
            "type": "string"
          },
          "challenge": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Web3NonceResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Web3NonceData"
          }
        }
      },
      "Web3VerifyRequest": {
        "type": "object",
        "required": [
          "wallet_address",
          "nonce",
          "signature"
        ],
        "properties": {
          "wallet_address": {
            "type": "string",
            "example": "0x1234567890abcdef1234567890abcdef12345678"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "login"
            ],
            "default": "login"
          },
          "nonce": {
            "type": "string"
          },
          "signature": {
            "type": "string",
            "description": "0x-prefixed secp256k1 signature from personal_sign"
          }
        }
      },
      "AuthUser": {
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "timezone",
          "locale",
          "plan"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "plan": {
            "type": "string"
          }
        }
      },
      "AuthSession": {
        "type": "object",
        "required": [
          "id",
          "status",
          "session_family_id",
          "last_seen_at",
          "expires_at",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "session_family_id": {
            "type": "string"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "revoked_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthLoginData": {
        "type": "object",
        "required": [
          "user",
          "session",
          "access_token",
          "access_token_expires_at",
          "refresh_token_expires_at",
          "token_type"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "session": {
            "$ref": "#/components/schemas/AuthSession"
          },
          "access_token": {
            "type": "string"
          },
          "access_token_expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "refresh_token_expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "token_type": {
            "type": "string"
          }
        }
      },
      "AuthLoginResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthLoginData"
          }
        }
      },
      "AuthMeResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "user",
              "session"
            ],
            "properties": {
              "user": {
                "$ref": "#/components/schemas/AuthUser"
              },
              "session": {
                "$ref": "#/components/schemas/AuthSession"
              }
            }
          }
        }
      },
      "AuthLogoutResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "ok"
            ],
            "properties": {
              "ok": {
                "type": "boolean",
                "const": true
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "not_found"
          },
          "message": {
            "type": "string",
            "example": "coin \"bitcoin\" not found"
          }
        }
      },
...
    }
  }
}